+++ load.py
@@ -1,32 +1,33 @@
 """
 A dummy plugin for testing and illustrative purposes.
 
-It doesn't do anything useful, but it does implement the plugin interface. It's used to enable unit testing of
+It doesn't do anything very useful, but it does implement the plugin interface. It's used to enable unit testing of
 the unit test harness making it a test harness test harness.
 
 It could also serve as a template for a new plugin.
-It stores the latest journal, dashboard, and carrier data in global variables for inspection.
+It displays journal events as they occur and stores the latest journal, dashboard, and carrier data in global variables for inspection.
 """
 import tkinter as tk
 from dataclasses import dataclass, field
 from typing import Dict
 from companion import CAPIData # type: ignore
 
-from utils.debug import Debug
-from utils.updater import Updater, read_version_file
+from demoplugin.utils.debug import Debug, catch_exceptions
+from demoplugin.utils.updater import Updater, read_version_file
+from demoplugin.ui import UI
 
-PLUGIN_NAME = "DummyPlugin"
+PLUGIN_NAME = "DemoPlugin"
 VERSION = "0.0.0" # placeholder -- plugin_start3() overwrites this
 
 GH_OWNER = "dwomble" # Github owner name
-GH_PROJECT = 'EDMC-DummyPlugin' #  Github project name
+GH_PROJECT = 'EDMC-DemoPlugin' #  Github project name
 
 @dataclass
 class plugin:
     plugin_dir:str = ""
     updater:Updater|None = None
     parent:tk.Frame|None = None
-    frame:tk.Frame|None = None
+    ui:UI|None = None
     closing:bool = False
 
 @dataclass
@@ -81,14 +82,16 @@ def plugin_stop():
 
     plugin.closing = True
 
-def plugin_app(parent):
+@catch_exceptions
+def plugin_app(parent:tk.Frame):
     """ Return a TK Frame for adding to the EDMC main window """
-    frame:tk.Frame = tk.Frame(parent)
     plugin.parent = parent
-    plugin.frame = frame
+    frame:tk.Frame = tk.Frame(parent)
+    plugin.ui = UI(frame)
+
     return frame
 
-def plugin_prefs(parent, cmdr: str, is_beta: bool):
+def plugin_prefs(parent:tk.Frame, cmdr: str, is_beta: bool):
     """ Return a TK Frame for adding to the EDMC settings dialog """
     prefs:tk.Frame = tk.Frame(parent)
     return prefs
@@ -107,12 +110,20 @@ def journal_entry(cmdr, is_beta, system, station, entry, state):
     journal.entry = entry
     journal.state = state
 
+    if not plugin.ui:
+        return
+
+    plugin.ui.add_entry(entry)
+
 def dashboard_entry(cmdr:str, is_beta:bool, entry:dict) -> None:
     """ Handle dashboard state changes """
 
     dashboard.cmdr = cmdr
     dashboard.is_beta = is_beta
     dashboard.entry = entry
+
+    if plugin.ui:
+        plugin.ui.update_dashboard(entry)
 
 def capi_fleetcarrier(data:CAPIData):
     """ Handle Fleet carrier data """