+++ CHANGELOG.md
@@ -4,11 +4,15 @@
 
 ### New Features
 
+* Made the demo plugin do something slightly useful, including a lightweight
+  status mode/pips/alert-badge summary and journal event list
+
 * Added an EDMC theme-aware ttk/tk object library (th) including:
   * TopLevel
   * Frame
   * LabelFrame
   * Label
+  * Text
   * Entry with placeholder text
   * Entry with autocompletion
   * Button
@@ -18,6 +22,7 @@
   * Listbox
   * Scale
   * Spinbox
+  * Progressbar
   * Tooltip
 
 ### Changes